requirements.txt
fastapi==0.112.2
httpx==0.27.2
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
faiss-cpu==1.8.0.post1
pdfplumber==0.11.4
python-dotenv==1.0.1
uvicorn==0.30.6
python-multipart